requirements.txt
numpy>=1.19.5
openai
pandas
python-dotenv
